requirements.txt
overpy==0.7
python-telegram-bot==21.0.1
PyYAML==6.0.1
dataclass-factory==2.16
redis~=5.0.3
python-dotenv==1.0.1